services/device/app/requirements.txt
annotated-types>=0.7.0
anyio>=4.7.0
certifi>=2024.8.30
click>=8.1.7
fastapi>=0.115.6
h11>=0.14.0
httpcore>=1.0.7
httpx>=0.28.1
idna>=3.10
paho-mqtt>=2.1.0
pydantic>=2.10.3
pydantic_core>=2.27.1
sniffio>=1.3.1
starlette>=0.41.3
typing_extensions>=4.12.2
uvicorn>=0.32.1
